Initial cleanup to remove global
