+++ src/schemas/LocalLanguageActions.ts
@@ -1,3 +1,5 @@
+import { Action } from "../typings";
+
 export const LocalLanguageActions: Record<string, Action[]> = {
   RPGLE: [
     {
